moved types and imports into separate file
