+++ di/container.ts
@@ -1,94 +1,11 @@
-import { Container, createContainer } from '@evyweb/ioctopus';
+import { Container, createContainer } from "@evyweb/ioctopus";
 
-import { IAuthenticationService } from '@/src/application/services/authentication.service.interface';
-import { ITransactionManagerService } from '@/src/application/services/transaction-manager.service.interface';
-import { IInstrumentationService } from '@/src/application/services/instrumentation.service.interface';
-import { ICrashReporterService } from '@/src/application/services/crash-reporter.service.interface';
-
-import { ITodosRepository } from '@/src/application/repositories/todos.repository.interface';
-import { IUsersRepository } from '@/src/application/repositories/users.repository.interface';
-
-import { ICreateTodoUseCase } from '@/src/application/use-cases/todos/create-todo.use-case';
-import { IDeleteTodoUseCase } from '@/src/application/use-cases/todos/delete-todo.use-case';
-import { IGetTodosForUserUseCase } from '@/src/application/use-cases/todos/get-todos-for-user.use-case';
-import { IToggleTodoUseCase } from '@/src/application/use-cases/todos/toggle-todo.use-case';
-import { ISignInUseCase } from '@/src/application/use-cases/auth/sign-in.use-case';
-import { ISignUpUseCase } from '@/src/application/use-cases/auth/sign-up.use-case';
-import { ISignOutUseCase } from '@/src/application/use-cases/auth/sign-out.use-case';
-
-import { ISignInController } from '@/src/interface-adapters/controllers/auth/sign-in.controller';
-import { ISignOutController } from '@/src/interface-adapters/controllers/auth/sign-out.controller';
-import { ISignUpController } from '@/src/interface-adapters/controllers/auth/sign-up.controller';
-import { IBulkUpdateController } from '@/src/interface-adapters/controllers/todos/bulk-update.controller';
-import { ICreateTodoController } from '@/src/interface-adapters/controllers/todos/create-todo.controller';
-import { IGetTodosForUserController } from '@/src/interface-adapters/controllers/todos/get-todos-for-user.controller';
-import { IToggleTodoController } from '@/src/interface-adapters/controllers/todos/toggle-todo.controller';
-
-import { registerMonitoringModule } from '@/di/modules/monitoring';
-import { registerAuthenticationModule } from '@/di/modules/authentication';
-import { registerTransactionManagerModule } from '@/di/modules/transaction-manager';
-import { registerTodosModule } from '@/di/modules/todos';
-import { registerUsersModule } from '@/di/modules/users';
-
-export const DI = {
-  // Services
-  IAuthenticationService: Symbol.for('IAuthenticationService'),
-  ITransactionManagerService: Symbol.for('ITransactionManagerService'),
-  IInstrumentationService: Symbol.for('IInstrumentationService'),
-  ICrashReporterService: Symbol.for('ICrashReporterService'),
-
-  // Repositories
-  ITodosRepository: Symbol.for('ITodosRepository'),
-  IUsersRepository: Symbol.for('IUsersRepository'),
-
-  // Use Cases
-  ICreateTodoUseCase: Symbol.for('ICreateTodoUseCase'),
-  IDeleteTodoUseCase: Symbol.for('IDeleteTodoUseCase'),
-  IGetTodosForUserUseCase: Symbol.for('IGetTodosForUserUseCase'),
-  IToggleTodoUseCase: Symbol.for('IToggleTodoUseCase'),
-  ISignInUseCase: Symbol.for('ISignInUseCase'),
-  ISignOutUseCase: Symbol.for('ISignOutUseCase'),
-  ISignUpUseCase: Symbol.for('ISignUpUseCase'),
-
-  // Controllers
-  ISignInController: Symbol.for('ISignInController'),
-  ISignOutController: Symbol.for('ISignOutController'),
-  ISignUpController: Symbol.for('ISignUpController'),
-  IBulkUpdateController: Symbol.for('IBulkUpdateController'),
-  ICreateTodoController: Symbol.for('ICreateTodoController'),
-  IGetTodosForUserController: Symbol.for('IGetTodosForUserController'),
-  IToggleTodoController: Symbol.for('IToggleTodoController'),
-};
-
-type DI_TYPES = {
-  // Services
-  IAuthenticationService: IAuthenticationService;
-  ITransactionManagerService: ITransactionManagerService;
-  IInstrumentationService: IInstrumentationService;
-  ICrashReporterService: ICrashReporterService;
-
-  // Repositories
-  ITodosRepository: ITodosRepository;
-  IUsersRepository: IUsersRepository;
-
-  // Use Cases
-  ICreateTodoUseCase: ICreateTodoUseCase;
-  IDeleteTodoUseCase: IDeleteTodoUseCase;
-  IGetTodosForUserUseCase: IGetTodosForUserUseCase;
-  IToggleTodoUseCase: IToggleTodoUseCase;
-  ISignInUseCase: ISignInUseCase;
-  ISignOutUseCase: ISignOutUseCase;
-  ISignUpUseCase: ISignUpUseCase;
-
-  // Controllers
-  ISignInController: ISignInController;
-  ISignOutController: ISignOutController;
-  ISignUpController: ISignUpController;
-  IBulkUpdateController: IBulkUpdateController;
-  ICreateTodoController: ICreateTodoController;
-  IGetTodosForUserController: IGetTodosForUserController;
-  IToggleTodoController: IToggleTodoController;
-};
+import { registerMonitoringModule } from "@/di/modules/monitoring";
+import { registerAuthenticationModule } from "@/di/modules/authentication";
+import { registerTransactionManagerModule } from "@/di/modules/transaction-manager";
+import { registerTodosModule } from "@/di/modules/todos";
+import { registerUsersModule } from "@/di/modules/users";
+import { DI_TYPES, DI } from "@/di/types";
 
 const container: Container = createContainer();
 
@@ -99,7 +16,7 @@ registerUsersModule(container);
 registerTodosModule(container);
 
 export function resolveDependency<T extends keyof DI_TYPES>(
-  key: T
+  key: T,
 ): DI_TYPES[T] {
   return container.get<DI_TYPES[T]>(DI[key]);
 }
